Migrate server command to new CLI
